pyproject.toml
[project]
name = "rf-serial-gui"
version = "0.1.0"
description = "RF Serial GUI Controller - Send values to RF transceiver via serial port"
requires-python = ">=3.8"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
rf-serial-gui = "rf_serial_gui.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rf_serial_gui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["src"]
